% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dirs.R
\name{build_dirs}
\alias{build_dirs}
\title{Build sub-directories for Shiny Server}
\usage{
build_dirs(yaml_loc = "/srv/shiny-server/app.yml")
}
\arguments{
\item{yaml_loc}{path to app.yml file}
}
\description{
This function builds sub-directories for apps in R packages and creates an
\code{app.R} file to launch each one on Shiny Server. This makes it easier to
install shiny apps via independent R packages and host them on Shiny Server.
}
\details{
A CI tool (e.g. Jenkins, Travis, or AppVeyor) can install the R package and
app on a server where Shiny Server is running. Then use \code{build_dirs()}
to update Shiny Server based on app.yml as part of the release step in a
CI/CD pipeline.
}
